requirements.txt
anthropic>=0.49.0
python-dotenv>=1.0.0
pytest>=8.0.0
ruff>=0.9.0
